test: add connectedCallback invocation timing check
