+++ packages/integration-karma/test/rendering/callback-invocation-order/x/lightContainer/lightContainer.js
