+++ packages/integration-karma/test/rendering/callback-invocation-order/x/lightParent/lightParent.js
