+++ packages/integration-karma/test/rendering/callback-invocation-order/x/lightShadowParent/lightShadowParent.js
