+++ packages/integration-karma/test/rendering/callback-invocation-order/x/parent/parent.js
