+++ packages/integration-karma/test/rendering/callback-invocation-order/x/shadowContainer/shadowContainer.js
