+++ packages/integration-karma/test/rendering/callback-invocation-order/x/shadowLightParent/shadowLightParent.js
