+++ packages/integration-karma/test/rendering/callback-invocation-order/x/shadowParent/shadowParent.js
